06-api-security-iam/requirements.txt
Flask>=2.3
PyJWT>=2.8